LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/responsive-iframe/responsive-iframe-navigation-expected.txt



FAIL Responsive iframe preserves size during navigation assert_equals: Initial height should be 200px expected 200 but got 150
